requirements.txt
torch==2.3.1
torchinfo==1.8.0
torchvision==0.18.1
pandas==2.2.2
pillow==10.3.0
numpy==1.23.0